[refactor/#43] Input 접근성 코드 정리
