+++ src/components/ui/Input/DateInput.tsx
@@ -1,9 +1,13 @@
 "use client";
 
 import DatePicker from "react-datepicker";
-import CalendarIcon from "@/assets/calender_icon.svg";
 import Image from "next/image";
+
+// Utils
 import { DateFormatToString } from "@/utils/DateFormatToString";
+
+// Icons
+import CalendarIcon from "@/assets/svgs/calender_icon.svg";
 
 type DateInputProps = {
   value: string | null;
@@ -15,7 +19,8 @@ export function DateInput({ value, onChange, placeholder }: DateInputProps) {
   return (
     <div className="flex flex-col gap-y-[10px]">
       <label className="text-xl text-gray-900">날짜</label>
-      <div className="relative w-full [&_.react-datepicker-wrapper]:w-full">
+      
+      <div className="relative w-[380px] [&_.react-datepicker-wrapper]:w-full">
         <DatePicker
           selected={value ? new Date(value) : null}
           onChange={(d: Date | null) =>
@@ -27,7 +32,8 @@ export function DateInput({ value, onChange, placeholder }: DateInputProps) {
         />
         <Image
           src={CalendarIcon}
-          alt="달력 아이콘"
+          alt=""
+          aria-hidden="true"
           width={32}
           height={32}
           className="absolute top-1/2 -translate-y-1/2 right-[12px]"
